🔄 (utils.py): refactor build_clean_params function to be shared between edge and vertex modules for code reusability and maintainability 📝 (utils.py): add build_clean_params function to vertex module to avoid circular import and improve code organization
